+++ Sources/TuistCore/Models/Project.swift
@@ -11,7 +11,6 @@ public struct Project: Equatable, CustomStringConvertible {
             lhs.targets == rhs.targets &&
             lhs.packages == rhs.packages &&
             lhs.schemes == rhs.schemes &&
-            lhs.autogenerateSchemes == rhs.autogenerateSchemes &&
             lhs.settings == rhs.settings &&
             lhs.filesGroup == rhs.filesGroup &&
             lhs.additionalFiles == rhs.additionalFiles
@@ -40,9 +39,6 @@ public struct Project: Equatable, CustomStringConvertible {
     /// Project schemes
     public var schemes: [Scheme]
 
-    /// Auto generate default schemes
-    public var autogenerateSchemes: Bool
-
     /// Project settings.
     public var settings: Settings
 
@@ -67,15 +63,14 @@ public struct Project: Equatable, CustomStringConvertible {
     ///                      *(Those won't be included in any build phases)*
     public init(path: AbsolutePath,
                 name: String,
-                organizationName: String? = nil,
-                fileName: String? = nil,
+                organizationName: String?,
+                fileName: String?,
                 settings: Settings,
                 filesGroup: ProjectGroup,
-                targets: [Target] = [],
-                packages: [Package] = [],
-                schemes: [Scheme] = [],
-                autogenerateSchemes: Bool = true,
-                additionalFiles: [FileElement] = []) {
+                targets: [Target],
+                packages: [Package],
+                schemes: [Scheme],
+                additionalFiles: [FileElement]) {
         self.path = path
         self.name = name
         self.organizationName = organizationName
@@ -83,7 +78,6 @@ public struct Project: Equatable, CustomStringConvertible {
         self.targets = targets
         self.packages = packages
         self.schemes = schemes
-        self.autogenerateSchemes = autogenerateSchemes
         self.settings = settings
         self.filesGroup = filesGroup
         self.additionalFiles = additionalFiles
@@ -144,7 +138,28 @@ public struct Project: Equatable, CustomStringConvertible {
                 targets: targets,
                 packages: packages,
                 schemes: schemes,
-                autogenerateSchemes: autogenerateSchemes,
                 additionalFiles: additionalFiles)
+    }
+
+    /// Returns a copy of the project with the given schemes set.
+    /// - Parameter schemes: Schemes to be set to the copy.
+    public func with(schemes: [Scheme]) -> Project {
+        Project(path: path,
+                name: name,
+                organizationName: organizationName,
+                fileName: fileName,
+                settings: settings,
+                filesGroup: filesGroup,
+                targets: targets,
+                packages: packages,
+                schemes: schemes,
+                additionalFiles: additionalFiles)
+    }
+
+    /// Returns the name of the default configuration.
+    public var defaultDebugBuildConfigurationName: String {
+        let debugConfiguration = settings.defaultDebugBuildConfiguration()
+        let buildConfiguration = debugConfiguration ?? settings.configurations.keys.first
+        return buildConfiguration?.name ?? BuildConfiguration.debug.name
     }
 }