+++ Sources/TuistKit/GraphMappers/GraphMapperProvider.swift
@@ -2,6 +2,7 @@ import Foundation
 import TuistCache
 import TuistCloud
 import TuistCore
+import TuistGenerator
 
 /// It defines an interface for providing the mappers to be used for a specific configuration.
 protocol GraphMapperProviding {