+++ Sources/TuistLoader/Models+ManifestMappers/Project+ManifestMapper.swift
@@ -21,6 +21,7 @@ extension TuistCore.Project {
         return Project(path: generatorPaths.manifestDirectory,
                        name: name,
                        organizationName: organizationName,
+                       fileName: nil,
                        settings: settings ?? .default,
                        filesGroup: .group(name: "Project"),
                        targets: targets,
@@ -39,7 +40,6 @@ extension TuistCore.Project {
                 targets: targets + [target],
                 packages: packages,
                 schemes: schemes,
-                autogenerateSchemes: autogenerateSchemes,
                 additionalFiles: additionalFiles)
     }
 
@@ -53,7 +53,6 @@ extension TuistCore.Project {
                 targets: targets,
                 packages: packages,
                 schemes: schemes,
-                autogenerateSchemes: autogenerateSchemes,
                 additionalFiles: additionalFiles)
     }
 
@@ -67,21 +66,6 @@ extension TuistCore.Project {
                 targets: targets,
                 packages: packages,
                 schemes: schemes,
-                autogenerateSchemes: autogenerateSchemes,
-                additionalFiles: additionalFiles)
-    }
-
-    func replacing(autogenerateSchemes: Bool) -> TuistCore.Project {
-        Project(path: path,
-                name: name,
-                organizationName: organizationName,
-                fileName: fileName,
-                settings: settings,
-                filesGroup: filesGroup,
-                targets: targets,
-                packages: packages,
-                schemes: schemes,
-                autogenerateSchemes: autogenerateSchemes,
                 additionalFiles: additionalFiles)
     }
 }